CRF: Renaming files according to linting standards
