+++ CameraProcessor/processor/dataloaders/json_dataloader.py
@@ -8,10 +8,10 @@ import json
 from processor.data_object.bounding_box import BoundingBox
 from processor.data_object.bounding_boxes import BoundingBoxes
 from processor.data_object.rectangle import Rectangle
-from processor.dataloaders.idataloader import IDataloader
+from processor.dataloaders.i_dataloader import IDataloader
 
 
-class JSONDataloader(IDataloader):
+class JsonDataloader(IDataloader):
     """JSON Dataloader, formats MOT Data."""
 
     def parse_file(self):
