+++ CameraProcessor/processor/dataloaders/mot_dataloader.py
@@ -10,10 +10,10 @@ from os import path
 from processor.data_object.bounding_box import BoundingBox
 from processor.data_object.bounding_boxes import BoundingBoxes
 from processor.data_object.rectangle import Rectangle
-from processor.dataloaders.idataloader import IDataloader
+from processor.dataloaders.i_dataloader import IDataloader
 
 
-class MOTDataloader(IDataloader):
+class MotDataloader(IDataloader):
     """MOT Dataloader, formats MOT Data."""
     def __init__(self, configs, path_location):
         """Initializes the MOTDataloader.
