+++ CameraProcessor/processor/input/cam_capture.py
@@ -9,7 +9,7 @@ import time
 import logging
 from cv2 import VideoCapture
 
-from processor.input.icapture import ICapture
+from processor.input.i_capture import ICapture
 from processor.data_object.frame_obj import FrameObj
