+++ CameraProcessor/processor/input/image_capture.py
@@ -10,7 +10,7 @@ import time
 import logging
 import cv2
 
-from processor.input.icapture import ICapture
+from processor.input.i_capture import ICapture
 from processor.data_object.frame_obj import FrameObj
