+++ CameraProcessor/processor/input/video_capture.py
@@ -8,7 +8,7 @@ Utrecht University within the Software Project course.
 import time
 import logging
 import cv2
-from processor.input.icapture import ICapture
+from processor.input.i_capture import ICapture
 from processor.data_object.frame_obj import FrameObj
