+++ CameraProcessor/processor/pipeline/detection/i_yolo_detector.py
@@ -7,7 +7,7 @@ Utrecht University within the Software Project course.
 import numpy as np
 import torch
 
-from processor.pipeline.detection.idetector import IDetector
+from processor.pipeline.detection.i_detector import IDetector
 from processor.data_object.bounding_box import BoundingBox
 from processor.data_object.rectangle import Rectangle
 from processor.pipeline.detection.yolor.utils.general import non_max_suppression, scale_coords
