+++ CameraProcessor/processor/pipeline/reidentification/fast_re_identifier.py
@@ -1,55 +1,58 @@
-"""Torch reid class.
+"""Fast reid class.
 
 This program has been developed by students from the bachelor Computer Science at
 Utrecht University within the Software Project course.
 © Copyright Utrecht University (Department of Information and Computing Sciences)
 """
 import os
+import argparse
 import copy
 import gdown
+from scipy.spatial.distance import cosine
 
-from scipy.spatial.distance import euclidean
-
-import processor.utils.features as UtilsFeatures
-from processor.pipeline.reidentification.ireidentifier import IReIdentifier
 from processor.data_object.bounding_box import BoundingBox
 from processor.data_object.bounding_boxes import BoundingBoxes
-from processor.pipeline.reidentification.torchreid.torchreid.utils import FeatureExtractor
+from processor.pipeline.reidentification.fastreid.fastreid.config import get_cfg
+from processor.pipeline.reidentification.fastreid.demo.predictor import FeatureExtractionDemo
+from processor.pipeline.reidentification.i_re_identifier import IReIdentifier
+import processor.utils.features as UtilsFeatures
 
 
-class TorchReIdentifier(IReIdentifier):
-    """Re-id class that uses torch-reid to extract and compare features.
+class FastReIdentifier(IReIdentifier):
+    """Re-id class that uses fast-reid to extract and compare features.
 
     Attributes:
-        extractor (FeatureExtractor): Extractor for the feature vectors.
+        extractor (FeatureExtractionDemo): Extractor for the feature vectors.
         config (configparser.SectionProxy): Re-ID configuration.
         threshold (float): Threshold from which a re-identification is included.
     """
 
     def __init__(self, config):
-        """Initialize torch re-identifier.
+        """Initialize fast re-identifier.
 
         Args:
             config (configparser.SectionProxy): Re-ID configuration.
         """
 
-        # The path where the model weight file should be located.
-        weights_path = os.path.join(config['weights_dir_path'], config['model_name'] + '.pth')
+        args = argparse.ArgumentParser(description="Feature extraction with reid models")
+        args.config_file = config['config_file_path']
+        args.parallel = config.getboolean('parallel')
+
+        # Load config from file and command-line arguments.
+        cfg = get_cfg()
+        cfg.merge_from_file(args.config_file)
+        cfg.freeze()
 
         # Download the weights if it's not in the directory.
         if not os.path.exists(config['weights_dir_path']):
             os.mkdir(config['weights_dir_path'])
 
-        if not os.path.exists(weights_path):
-            url = 'https://drive.google.com/u/0/uc?id=1vduhq5DpN2q1g4fYEZfPI17MJeh9qyrA&export=download'
-            output = weights_path
+        if not os.path.exists(cfg.MODEL.WEIGHTS):
+            url = 'https://github.com/JDAI-CV/fast-reid/releases/download/v0.1.1/market_sbs_R101-ibn.pth'
+            output = cfg.MODEL.WEIGHTS
             gdown.download(url, output, quiet=False)
 
-        # Initialize the feature extractor of torch re-id.
-        self.extractor = FeatureExtractor(
-            model_name=config['model_name'],
-            model_path=weights_path,
-            device=config['device'])
+        self.extractor = FeatureExtractionDemo(cfg, parallel=args.parallel)
 
         self.config = config
         self.threshold = float(self.config["threshold"])
@@ -72,23 +75,9 @@ class TorchReIdentifier(IReIdentifier):
         resized_cutout = UtilsFeatures.resize_cutout(cutout, self.config)
 
         # Extract the feature from the cutout and convert it to a normal float array.
-        feature = self.extractor(resized_cutout).cpu().numpy().tolist()[0]
+        feature = self.extractor.run_on_image(resized_cutout).cpu().numpy().tolist()
 
         return feature
-
-    def extract_features_from_cutout(self, cutout):
-        """Extracts features from a cutout.
-
-        Args:
-            cutout (np.ndarray): the cutout containing the object we want to extract features from
-
-        Returns:
-            [float]: Feature vector of the cutout
-        """
-        # Resize the cutout.
-        resized_cutout = UtilsFeatures.resize_cutout(cutout, self.config)
-
-        return self.extractor(resized_cutout).cpu().numpy().tolist()[0]
 
     def extract_features_boxes(self, frame_obj, boxes):
         """Extracts features from all bounding boxes generated in the tracking stage.
@@ -111,8 +100,8 @@ class TorchReIdentifier(IReIdentifier):
         """Calculates the similarity rate between two feature vectors.
 
         Note:
-            Uses euclidean distance to determine the similarity.
-            An alternative would be cosine similarity.
+            Uses cosine similarity to determine the similarity rate.
+            An alternative would be euclidean distance.
 
         Args:
             query_features ([float]): the feature vector of the query image.
@@ -121,11 +110,11 @@ class TorchReIdentifier(IReIdentifier):
         Returns:
             float: The similarity value of two feature vectors.
         """
-        euclidean_distance = euclidean(query_features, gallery_features)
-        return euclidean_distance
+        cosine_similarity = 1 - cosine(query_features, gallery_features)
+        return cosine_similarity
 
     def re_identify(self, frame_obj, track_obj, re_id_data):
-        """Performing re-identification using Torchreid.
+        """Performing re-identification using torchreid.
 
         This re-identification implementations couple bounding boxes to a tracked subject
         which is not currently detected on the camera. Updates list of bounding box by possibly assigning an object ID
@@ -159,7 +148,7 @@ class TorchReIdentifier(IReIdentifier):
                 if tracked_bounding_boxes[i].object_id is None:
                     # Calculate the similarity value of the 2 feature vectors.
                     similarity_value = self.similarity(query_feature, feature)
-                    if similarity_value < self.threshold:
+                    if similarity_value > self.threshold:
                         box_id = tracked_bounding_boxes[i].identifier
 
                         # Store that this box id belongs to a certain object id.
