+++ CameraProcessor/processor/pipeline/tracking/i_sort_tracker.py
@@ -7,7 +7,7 @@ Utrecht University within the Software Project course.
 import numpy as np
 
 from processor.data_object.bounding_box import BoundingBox
-from processor.pipeline.tracking.itracker import ITracker
+from processor.pipeline.tracking.i_tracker import ITracker
 from processor.data_object.bounding_boxes import BoundingBoxes
 from processor.data_object.rectangle import Rectangle
