+++ CameraProcessor/processor/scheduling/component/i_component.py
