+++ CameraProcessor/processor/scheduling/node/i_node.py
