+++ CameraProcessor/processor/utils/create_runners.py
@@ -5,12 +5,12 @@ Utrecht University within the Software Project course.
 © Copyright Utrecht University (Department of Information and Computing Sciences)
 """
 
-from processor.pipeline.detection.yolov5_runner import Yolov5Detector
-from processor.pipeline.detection.yolor_runner import YolorDetector
+from processor.pipeline.detection.yolov5_detector import Yolov5Detector
+from processor.pipeline.detection.yolor_detector import YolorDetector
 from processor.pipeline.tracking.sort_tracker import SortTracker
-from processor.pipeline.tracking.sort_oh_tracker import SortOHTracker
-from processor.pipeline.reidentification.torchreid_runner import TorchReIdentifier
-from processor.pipeline.reidentification.fastreid_runner import FastReIdentifier
+from processor.pipeline.tracking.sort_oh_tracker import SortOhTracker
+from processor.pipeline.reidentification.torch_re_identifier import TorchReIdentifier
+from processor.pipeline.reidentification.fast_re_identifier import FastReIdentifier
 
 DETECTOR_SWITCH = {
     'yolov5': (Yolov5Detector, 'Yolov5'),
@@ -18,7 +18,7 @@ DETECTOR_SWITCH = {
 }
 TRACKER_SWITCH = {
     'sort': (SortTracker, 'SORT'),
-    'sort_oh': (SortOHTracker, 'SORT')
+    'sort_oh': (SortOhTracker, 'SORT')
 }
 REID_SWITCH = {
     'torchreid': (TorchReIdentifier, 'TorchReid'),
