+++ CameraProcessor/tests/unittests/pipeline/test_process_frames.py
@@ -9,12 +9,12 @@ import pytest
 
 from tests.unittests.utils.fake_detector import FakeDetector
 from tests.unittests.utils.fake_tracker import FakeTracker
-from tests.unittests.utils.fake_reidentifier import FakeReIdentifier
+from tests.unittests.utils.fake_re_identifier import FakeReIdentifier
 from tests.unittests.utils.fake_websocket import FakeWebsocket
 from processor.pipeline.prepare_pipeline import prepare_objects
 from processor.pipeline.process_frames import process_stream
-from processor.pipeline.detection.yolov5_runner import Yolov5Detector
-from processor.pipeline.detection.yolor_runner import YolorDetector
+from processor.pipeline.detection.yolov5_detector import Yolov5Detector
+from processor.pipeline.detection.yolor_detector import YolorDetector
 from processor.input.video_capture import VideoCapture
 from processor.pipeline.process_frames import send_boxes_to_orchestrator
