+++ CameraProcessor/tests/unittests/training/test_text_parsing.py
