+++ CameraProcessor/tests/unittests/utils/fake_tracker.py
@@ -10,7 +10,7 @@ import random
 from processor.data_object.rectangle import Rectangle
 from processor.data_object.bounding_box import BoundingBox
 from processor.data_object.bounding_boxes import BoundingBoxes
-from processor.pipeline.tracking.itracker import ITracker
+from processor.pipeline.tracking.i_tracker import ITracker
 
 
 class FakeTracker(ITracker):
