+++ CameraProcessor/tests/unittests/utils/test_features.py
@@ -1,4 +1,4 @@
-"""Tests the reid utils.
+"""Tests the utils.features.py.
 
 This program has been developed by students from the bachelor Computer Science at
 Utrecht University within the Software Project course.
@@ -12,7 +12,7 @@ from tests.unittests.utils.conftest import X0, X1, Y0, Y1
 from processor.utils.features import slice_bounding_box, resize_cutout
 
 
-class TestReidUtils:
+class TestFeatures:
     """Class to hold functions to test reid utils."""
     def test_slice_bounding_box(self, img, bbox):
         """Tests the slice_bounding_box method.
